Set up React app for handling front-end display
I'm using React to handle this project's front-end display because it is lightweight and can be quickly built and optimized for production. I also considered Angular, but I want to focus on encrypted communication in this project and I find Angular to be too dependent on internals which I am not yet familiar with, which might be distracting here. I am generally not familiar enough with other front-end frameworks to want to spend the time learning them now for this project. So, the simplicity of the React library wins for now so that I can focus on cryptography and still make a production build optimized for the modern web.