Add a unit test for encrypting and decrypting text
The Cipher class should have two instance methods: #encrypt(text) and #decrypt(text). The return type of each method should be a string. When #encrypt(text) encrypts an input string, the output string should be different from the input string. When the output of #encrypt(text) is decrypted by #decrypt(text), the output string should be equivalent to the original input string that was passed into #encrypt(text).