Add type checking tests for Cipher methods
Cipher#encrypt(text) should not throw a TypeError if a String is passed in. Cipher#decrypt(text) should not throw a TypeError if an encrypted String is passed in, but should throw a TypeError if an unencrypted String is passed in. Cipher methods should throw a TypeError if data is passed in that is of type Number, Object, Boolean, Null, or Undefined.