Implement encryption and decryption methods
Cipher#encrypt(text) takes a string and returns an encrypted string. Cipher#decrypt(text) takes an encrypted string and returns the original string that it was encrypted from.

These methods require a password within the scope of the Cipher class. This password is coming from a gitignored secrets file for now.