Add a unit test for signing an HTTP request
This test verifies that HTTPClient#sign_request returns an object containing an XMLHttpRequest and a digital signature.

As described in some comments in the test, an XMLHttpRequest object has no method for getting the values of request headers once they have been set, so I have not yet figured out how to test that the digital signature is actually set as the value of the request's Authorization header. My solution to this problem involves the following:
  * Testing that HTTPClient#sign_request returns an object containing an XMLHttpRequest
  * Testing that HTTPClient#sign_request returns an object containing a string, which is the digital signature
  * Leaving a console.log in the test that explains how to visually check that the XMLHttpRequest has an Authorization header with a string as its value
  * Leaving a comment above the line of code where the Authorization header is set, which explains to never delete that line of code since the tests will still pass but the request will be unsigned