Test the digital signature's exact value
This commit makes the digital signature more explicitly testable. A test password can now be set in the unit tests that results in a predictable digital signature. The tests now verify the exact value of the digital signature, and also verify that the decrypted digital signature is equivalent to the HTTP request line.

To achieve this, this commit changes the Cipher constructor to optionally allow a certain password to be set when creating a new cipher (instead of necessarily getting the password from a version-control-ignored secrets file). Also, the HTTPClient constructor now takes a cipher as an argument, instead of creating the cipher inside the #sign_request method. The method signature of HTTPClient#sign_request also now includes all of the pieces of the request line as arguments.