Use an actual API endpoint in the tests
I am in the process of building the back-end API. In order to use a consistent set of valid test data, I am replacing the fake API endpoint in the client-side tests with an actual API endpoint that the back-end will serve up. This results in a combination of a password, API endpoint, and digital signature that I can use as valid data in both the client-side and server-side tests.