Use an AES initialization vector when creating a Cipher
I have been having difficulty getting the server-side Ruby OpenSSL Advanced Encryption Standard wrapper to decrypt an encrypted message from this client-side code. I am investigating whether this problem is because the Ruby implementation requires an initialization vector (IV), whereas this client-side code was not using an IV beforehand. This commit includes changes that generate a random IV which is passed into a new Cipher along with a key. A possible next step is to try to use the same combination of key and IV to encrypt/create a signature on the client-side and decrypt it on the server-side.