Transmit the initialization vector as an HTTP header
The server will need to use the AES initialization vector (IV) to decrypt the digital signature. Since the IV is randomly generated, and is OK to be publicly accessible, this commit includes the IV as a string in an HTTP header so that the server can use it for decryption.