Update README.md
This commit updates the README to indicate that the authentication scheme will use a SHA-256 HMAC of the request line as the digital signature, instead of an AES encryption of the request line. This change is because I have learned that HMAC hashing is a more conventional way of generating digital signatures.