Add a test for generating an HMAC
I have learned that a more conventional format for a digital signature is a Hash-based Message Authentication Code, versus the AES encrypted string that I was working on earlier. So, I'm pivoting to set u
p the digital signature as a SHA-256 HMAC hash of the HTTP request line. This commit adds a test for generating an HMAC which will later be used for generating a digital signature.