Add a test for generating a digital signature
This commit adds a test which checks that HTTPClient#generate_signature(httpMethod, url, httpVersion, key) generates a digital signature which is the SHA-256 HMAC of the HTTP request line. It also makes a minor syntactical change to the test for HMAC#hash(data, key) by setting the expected hash as a variable that is compared to the actual hash.

It also removes a test for some prior functionality which I will remove. The method HTTPClient#sign_request(xhr, httpMethod, url, httpVersion), which took an XHR object and HTTP parameters and added a digital signature to the XHR, will be removed. It will be replaced by HTTPClient#generate_signature(httpMethod, url, httpVersion, key), which will only generate a digital signature and not add it as a header to an XHR object.

The intention is to make the digital signature more adherent to web conventions by using SHA-256 HMAC to generate it, and also to make the signature generation method more testable and more adherent to the single responsibility principle by only assigning it the job of generating a signature.