Implement generating a digital signature
This commit implements HTTPClient#generate_signature(httpMethod, url, httpVersion, key), which takes four strings as arguments and returns a string. The return value is a digital signature which is the SHA-256 HMAC of the HTTP request line.

This commit removes HTTPClient#sign_request(xhr, httpMethod, url, httpVersion), which not only generated a signature but also added it as a header on an XHR object. That approach did multiple things in one method and was hard to test. The new approach, where the method for generating a signature only generates the signature and does nothing else, is more testable and is more in line with the single responsibility principle.