Implement decrypting an encrypted HTTP response
This commit implements the method HTTPClient#decryptHTTP(response). This method takes a JSON object representing the HTTP response.

When the API responds with a body containing an encrypted message, as well as a header containing the AES initialization vector that was used to encrypt the message, the HTTP Client should be able to decrypt the message.