Receive the AES initialization vector in the response body
This commit applies changes that match an improvement in the API architecture which allows requests for multiple resources. Previously, in the API for getting a single resource, the initialization vector of that resource's encryption was passed in the HTTP response via a header. That architecture would not work in a request for several resources, because a different initialization vector is used to encrypt each resource's message. By passing the initialization vector in JSON in the response body, it will be possible to keep a consistent architecture for requests for both single and multiple resources, while still associating each encrypted message with its own initialization vector.