Update tests for a consistent response format
This commit applies changes to the tests that specify a consistent format for HTTP responses, regardless of whether the response contains a single resource or multiple resources. This should simplify the logic needed to parse the data in the response.

The client will need to parse an HTTP response containing a top-level JSON object with the key "messages", where the corresponding value is an array of objects that each contain an encrypted message and the AES initialization vector used to do the encryption.