Implement decrypting multiple messages in a response
This commit implements decrypting a response that contains several encrypted messages.

As I was working on this, I realized that I had set up inaccurate test data (it was not based on the key that the httpClient was using). To ensure that this problem doesn't happen again, I added a comment in the httpClient test that specifies what key to use and where to put it. It belongs in a secrets file that is ignored from version control, and under the current architecture I have not figured out how to override it explicitly in the tests with a unique test key. So, my solution for now is specifying the test key in the comment at the top of the test file, and keeping all of the test data consistent with that key.