Add a comment explaining test setup for Cipher unit tests
During my work on my previous commit, I realized that I had not provided enough documentation on how to set up the test environment. This commit adds a comment to the unit tests for the Cipher class, which explains how to set up a random initialization vector that is used by default whenever an initialization vector is not explicitly passed in to the Cipher constructor.

Also, this commit removes a couple of leftover debugging comments in the httpClient class.