Add tests for authenticating an HTTP response
This commit adds tests for the method httpClient#authenticate(response). This method returns true if every signature in the response matches a valid signature which is generated by decrypting the associated message cipher and hashing the unencrypted message text with SHA-256 HMAC, using a secret key that is agreed upon by the client and the server. This method returns false if there is at least one signature in the response that does not match the valid signature generated by the client.

This method enables the client to verify the origin and integrity of the response.

Also, this commit adds an additional test covering the case of a response with one valid signature and one invalid signature, which should not be authenticated.